Update updateOmbi.sh
I'm sure there is a better way of doing this. GitHub does not show the 'Assets' portion automatically anymore on the releases page. This was the easiest way I could think of to grab the latest version number while leaving the rest of the code as is because it still works (the updating based off the old URL).